requirements.txt
apache-airflow
dbt-core
dbt-duckdb
sling
python-dotenv